feat: refacto app and main layouts (#64)
