+++ src/app/[locale]/layout.tsx
@@ -1,34 +1,24 @@
-import type { Metadata } from "next";
+import React from "react";
 import "../../../assets/sass/core.scss";
-import { merriweather_Sans } from "../../utils/font";
-import FooterComponent from "src/components/layout/footer.component";
-import { getMessages } from "next-intl/server";
-import { NextIntlClientProvider } from "next-intl";
 import HeaderComponent from "src/components/layout/header/header.component";
+import FooterComponent from "src/components/layout/footer.component";
+import { NextIntlClientProvider } from "next-intl";
+import { getMessages } from "next-intl/server";
 
-export const metadata: Metadata = {
-  title: "Le restaurant - Restaurant Van Long",
-  description:
-    "Situé dans le 13ème arrondissement de Paris depuis 1992,de la rue Pascal à la rue du Champ de l’Alouette,notre restaurant traditionnel vietnamien vous accueille dans un cadre simple et chaleureux. Tous nos plats sont élaborés sur place, à partir de produits frais, et vous rappelleront les saveurs subtiles de la cuisine vietnamienne. Précédent Suivant",
+export type TMainLayout = {
+  children: React.ReactNode;
 };
 
-export default async function RootLayout({
-  children,
-  params: { locale },
-}: Readonly<{
-  children: React.ReactNode;
-  params: { locale: string };
-}>) {
+const MainLayout = async ({ children }: TMainLayout) => {
   const messages = await getMessages();
+
   return (
-    <html lang={locale} className={merriweather_Sans.className}>
-      <body>
-        <NextIntlClientProvider messages={messages}>
-          <HeaderComponent />
-          {children}
-          <FooterComponent />
-        </NextIntlClientProvider>
-      </body>
-    </html>
+    <NextIntlClientProvider messages={messages}>
+      <HeaderComponent />
+      {children}
+      <FooterComponent />
+    </NextIntlClientProvider>
   );
-}
+};
+
+export default MainLayout;
